Make the video server info JSON have ephemeral liveness
We're about to use it to speed up server restarts, so it needs to
have only minimal caching.

I've still set it to have *some* caching so that we don't end up
with the server being DoS'd. Really, 1 s expiry should be fine.

I'm not sure why CF is serving hours old segments though.