Remove AV interleaving tweaks from the video server
They don't actually appear to be necessary, but they are
interfering with AV sync.

Audio has a shorter encoding latency than video, so the audio is
available first. But we told it the deinterleaving latency had to
be small AND it had to flush packets immediately. I guess that
forced it to desynchronize the audio and video. Oops.