Improve handling of body lengths
We don't need to explicitly call `readEmpty()` if we don't want a
request body. We also now correctly reject absurdly long request
bodies without loading them all into memory.

This removes the broken all-read check from ~TestResource, too.
The server tests that requests are fully consumed (correctly), so
we should have tests that test the correctness of that, and then
rely on it to throw during other tests. It's really silly to
reimplement portions of the application's logic (in slightly
different ways) in tests like that. Just test that the app does
the thing it's supposed to, and then rely on it in later tests.