Make liburing optional
Unfortunately, there seems to be a bug in boost::process::v2 that
means that the uring event loop can get stuck if a subprocess is
created while other IO operations are happening.

Unfortunately, to remove the uring dependency, I had to use
synchronous IO for files, because epoll doesn't support them. An
enhancement on this would be to use a thread for the file IO, and
have a notification system to synchronize with the main thread.