Implement CORS preflight support
So the CORS stuff actually works

This implementation is not ideal, and is slightly insecure, because
I can't be arsed to make the Request objects suck less. Why is it
split into two classes? This looks like horrible structure just to
cater to the shitt tests?

It's impossible to access request headers from request handlers,
it seems? WTF?