[Server] Better enforcement of segment boundaries
I'm not entirely certain why they were not being properly enforced
before. Zac exhibited a video which appeared to have the property
that scene-cut I frames were causing the segments start I-frames to
be in the wrong place. I'm not sure why, but this fixes it (and
also allows for a slight drift in segment start time with frame
rates that are not a multiple of the segment duration to be
corrected with segment GOPs that vary between two values).