Don't poll the channel index madly
The channel index is expected to change very rarely. User code should
probably govern when it gets changed.

We could perhaps implement a server-mediated "refresh configuration"
signal that would trigger it, among other things. Polling is not
sensible: it was accounting for the majority of the http calls
made by the program - and that shit ain't free :D